refactor: split db client out of index re-export
